graphql-service/requirements.txt
strawberry-graphql
sqlalchemy
starlette
psycopg2-binary
uvicorn
python-dotenv